First attempt for toast that shows on SW update Housekeeping And other stuff
