+++ gatsby-browser.js
@@ -10,7 +10,7 @@ const { getLocale, getLanguage } = require('./src/utils/locale')
 const printCorporateMessage = require('./gatsby/browser/corporateMessage')
 const setDefaultTime = require('./gatsby/browser/defaultTime')
 
-require('./src/styles/app.scss')
+require('./src/styles/app.scss') // main styling
 
 // module.exports.wrapPageElement = require('./gatsby/wrapPageElement')
 module.exports.wrapRootElement = require('./gatsby/wrapRootElement')
@@ -28,7 +28,7 @@ module.exports.onInitialClientRender = () => {
   printCorporateMessage()
 }
 
-// exports.onRouteUpdate = ({ location }) => {
-//   const path = location.pathname + location.search + location.hash
-//   console.info('Track pageview of:', path)
-// }
+module.exports.onServiceWorkerUpdateReady = () => {
+  console.info('NEW VERSION AVAILABLE')
+  // window.location.reload()
+}
