+++ src/components/Breadcrumb.jsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react'
 import { Breadcrumb as BreadcrumbPlugin } from 'gatsby-plugin-breadcrumb'
 
-import { HistoryContext } from '../store/history'
+import { HistoryContext } from '../provider/history'
 
 const Breadcrumb = () => {
   const { location, crumbLabel, crumbs } = useContext(HistoryContext)
