+++ src/styles/_breakpoints.scss
@@ -1,7 +1,6 @@
 // Destinations to check, after editing `$grid-breakpoints`:
 // src/utils/responsive.js
-// src/fragments/sharp.js
-// gatsby/config/commonRemark.js
+// gatsby-config.js#gatsby-remark-images
 
 $grid-breakpoints: (
   xs: 0,
