+++ src/styles/app.scss
@@ -47,6 +47,7 @@
 @import 'components/language';
 @import 'components/logo';
 @import 'components/my-age';
+@import 'components/no-script';
 @import 'components/overlay';
 @import 'components/portfolio';
 @import 'components/quote';
@@ -54,6 +55,7 @@
 @import 'components/separator';
 @import 'components/status';
 @import 'components/theme-switch';
+@import 'components/toast';
 @import 'components/tooltip';
 
 // ------------------------------
