+++ src/styles/components/_button.scss
@@ -10,10 +10,11 @@ button {
 }
 
 .btn {
-  display: inline-block;
   position: relative;
   padding: 10px 18px;
   transition: color 200ms ease-in-out;
+  border: 0;
+  background: transparent;
   color: var(--text-color-normal);
   font-size: var(--text-sm);
   font-weight: var(--font-bold);
@@ -24,5 +25,16 @@ button {
 
   &:hover {
     color: var(--text-color-richer);
+  }
+
+  &--small {
+    padding: 6px 12px;
+    font-size: var(--text-xs);
+
+    @include border-left-right(10px);
+  }
+
+  &--inline {
+    display: inline-block;
   }
 }
