+++ src/styles/components/_no-script.scss
@@ -1,0 +1,4 @@
+#gatsby-noscript {
+  position: fixed;
+  font-size: var(--text-xxl);
+}
