+++ src/styles/components/_toast.scss
@@ -1,0 +1,44 @@
+.toasts-wrapper {
+  display: flex;
+  position: fixed;
+  bottom: 0;
+  left: 0px;
+  flex-direction: column;
+  margin: var(--spacing-sm);
+}
+
+.toast {
+  position: relative;
+  width: 300px;
+  max-width: 100%;
+  margin-top: var(--spacing-sm);
+  padding: var(--spacing-sm);
+  background: var(--color-primary);
+  color: var(--color-dark);
+
+  &__text {
+    font-size: var(--text-md);
+  }
+
+  &__actions {
+    display: flex;
+    margin-top: var(--spacing-sm);
+    column-gap: var(--spacing-xs);
+
+    button {
+      flex: 1;
+      color: var(--text-color-dark);
+    }
+  }
+
+  &__btn-close {
+    display: flex;
+    position: absolute;
+    top: var(--spacing-xs);
+    right: var(--spacing-xs);
+
+    .icon {
+      --icon-size: 12px;
+    }
+  }
+}
