+++ src/sw.js
@@ -1,0 +1,2 @@
+/* globals workbox */
+/* eslint-disable no-restricted-globals */
