--- src/utils/transform.js
@@ -1,3 +1,0 @@
-export function capitalizeString(string) {
-  return string.charAt(0).toUpperCase() + string.slice(1)
-}
